feat: add slug support for courses across all components
