+++ backend/courses/views.py
@@ -1,4 +1,5 @@
 from rest_framework import viewsets, permissions, status
+from rest_framework.exceptions import APIException, NotFound
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from django.http import HttpResponse
@@ -33,6 +34,35 @@ class IsAdminUserOrReadOnly(permissions.BasePermission):
 
 class CourseViewSet(viewsets.ModelViewSet):
     queryset = Course.objects.all()
+    lookup_field = 'slug'
+
+    def get_object(self):
+        """Resolve object by slug first, then fall back to numeric PK if needed.
+
+        This allows older clients that still send numeric IDs in the URL to work
+        while the viewset primarily uses slug lookups.
+        """
+        lookup_value = self.kwargs.get(self.lookup_field) or self.kwargs.get('pk')
+        queryset = self.filter_queryset(self.get_queryset())
+
+        # Try slug lookup first
+        if lookup_value is None:
+            raise NotFound(detail="Course not found")
+
+        try:
+            obj = queryset.get(**{self.lookup_field: lookup_value})
+            self.check_object_permissions(self.request, obj)
+            return obj
+        except Course.DoesNotExist:
+            # If the lookup value looks numeric, try PK fallback
+            if str(lookup_value).isdigit():
+                try:
+                    obj = queryset.get(pk=lookup_value)
+                    self.check_object_permissions(self.request, obj)
+                    return obj
+                except Course.DoesNotExist:
+                    pass
+            raise NotFound(detail="Course not found")
 
     def get_queryset(self):
         qs = Course.objects.all()
@@ -51,7 +81,12 @@ class CourseViewSet(viewsets.ModelViewSet):
             # The file deletion is handled in the model's delete method
             self.perform_destroy(instance)
             return Response(status=status.HTTP_204_NO_CONTENT)
+        except APIException:
+            # Let DRF handle expected API exceptions (404, 403, etc.)
+            raise
         except Exception:
+            # Log unexpected exceptions and return 500
+            logging.exception("Error while deleting course")
             return Response(
                 {"detail": "An error occurred while deleting the course."},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR
@@ -61,7 +96,11 @@ class CourseViewSet(viewsets.ModelViewSet):
         """Override update method to handle file replacement properly."""
         try:
             return super().update(request, *args, **kwargs)
+        except APIException:
+            # Re-raise DRF API exceptions so they are converted to proper 4xx responses
+            raise
         except Exception:
+            logging.exception("Error while updating course")
             return Response(
                 {"detail": "An error occurred while updating the course."},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR
@@ -71,7 +110,10 @@ class CourseViewSet(viewsets.ModelViewSet):
         """Override partial_update method to handle file replacement properly."""
         try:
             return super().partial_update(request, *args, **kwargs)
+        except APIException:
+            raise
         except Exception:
+            logging.exception("Error while partially updating course")
             return Response(
                 {"detail": "An error occurred while updating the course."},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR
@@ -194,7 +236,8 @@ class CourseViewSet(viewsets.ModelViewSet):
             # print(f"[Stripe Checkout] Session created: {session.id}")
             return Response({"checkout_url": session.url})
         except Exception:
-            return Response({"detail": "An internal error occurred during Stripe checkout."}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            return Response({"detail": "An internal error occurred during Stripe checkout."},
+                            status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
     @action(detail=True, methods=['post'], permission_classes=[permissions.IsAuthenticated], url_path='refund-course')
     def refund_course(self, request, pk=None):
@@ -226,7 +269,7 @@ class CourseViewSet(viewsets.ModelViewSet):
                             status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
         try:
-            refund = stripe.Refund.create(payment_intent=enrollment.stripe_payment_intent_id)
+            stripe.Refund.create(payment_intent=enrollment.stripe_payment_intent_id)
             enrollment.delete()
             return Response({"detail": "Refund processed and unenrolled from course."})
         except Exception as e:
