+++ frontend/src/components/admin/admin-course-modal.tsx
@@ -7,6 +7,7 @@ import EnrolledMembers from "@/components/admin/enrolled-members";
 
 interface Course {
   id: number;
+  slug?: string;
   title: string;
   subtitle?: string;
   description: string;
